ExcelImportModal changed to ExcelImportSideSheet. General look an functionality is the same.
