+++ src/frontend/src/pages/ProjectPage/pages/ContractPage/pages/ManagePersonnelPage/components/ExcelImportSideSheet/styles.less
@@ -2,14 +2,10 @@
 
 .excelImportModal {
   display:flex;
-  padding:0;
-  margin: 0 auto;
-  margin-top: 20%;
-  width: 500px;
-  background-color:var(--color-white);
+  width: 100%;
+  height: 30%;  
   justify-content: center;
   justify-items: center;
-  border-radius: 4px ;
   
   .inputs {
     margin: 5px;
@@ -26,7 +22,9 @@
       border: dashed;
       border-radius: 4px;
       border-color: var(--color-primary);
+      height: 80%;
       width: 100%;
+
       
       .dragDropText {
         text-align: center;
